% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two.R
\encoding{UTF-8}
\name{add_two}
\alias{add_two}
\title{숫자 두개를 더해서 결과를 반환시키는 프로그램
Add two numbers and retun added summation}
\usage{
add_two(a, b)
}
\arguments{
\item{a}{numeric scalar}

\item{b}{numeric scalar}
}
\description{
숫자 두개를 더해서 결과를 반환시키는 프로그램
Add two numbers and retun added summation
}
