% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{c2e}
\alias{c2e}
\title{Character to Expression}
\usage{
c2e(x)
}
\arguments{
\item{x}{characters}
}
\value{
expressions.
}
\description{
Converts characters to expressions
}
\examples{
# convert characters
c2e('a')

# convert array of characters
c2e(array('a', dim = c(2,2)))

}
\seealso{
\code{\link{e2c}}
}
