% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{calculus-package}
\alias{calculus}
\alias{calculus-package}
\title{calculus: High Dimensional Numerical and Symbolic Calculus}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Efficient C++ optimized functions for numerical and symbolic calculus as described in Guidotti (2020) \href{https://arxiv.org/abs/2101.00086}{arXiv:2101.00086}. It includes basic arithmetic, tensor calculus, Einstein summing convention, fast computation of the Levi-Civita symbol and generalized Kronecker delta, Taylor series expansion, multivariate Hermite polynomials, high-order derivatives, ordinary differential equations, differential operators (Gradient, Jacobian, Hessian, Divergence, Curl, Laplacian) and numerical integration in arbitrary orthogonal coordinate systems: cartesian, polar, spherical, cylindrical, parabolic or user defined by custom scale factors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://calculus.guidotti.dev}
  \item Report bugs at \url{https://github.com/eguidotti/calculus/issues}
}

}
\author{
\strong{Maintainer}: Emanuele Guidotti \email{emanuele.guidotti@unine.ch} (\href{https://orcid.org/0000-0002-8961-6623}{ORCID})

}
\keyword{internal}
