% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derivative & definiteIntegral.R
\name{definiteIntegral}
\alias{definiteIntegral}
\title{Definite Integral function}
\usage{
definiteIntegral(expression = NULL, start, end, variable = "X")
}
\arguments{
\item{expression}{A string contatining the function to integrate.}

\item{start}\item{end}{Limits of integration}

\item{variable}{A string containing the variable used in @param expression, Make sure the variable
string does not show anywhere else in @param expression (Ex: variable "x" in "exp(x)") To avoid this,
defaults to "X".}
}
\description{
Calculates the definite integral of a function between two given points.
}
\examples{
definiteIntegral("sin(X)*exp(X)",0,12)
}
