% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det.R
\name{det}
\alias{det}
\title{Numerical and Symbolic Determinant}
\usage{
det(x)
}
\arguments{
\item{x}{numeric or character matrix.}
}
\value{
numeric or character determinant.
}
\description{
Calculates the determinant of a matrix.
}
\examples{
# numeric matrix
x <- matrix(1:4, nrow = 2)
det(x)

# symbolic matrix
x <- matrix(letters[1:4], nrow = 2)
det(x)

}
