% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{divergence}
\alias{divergence}
\alias{\%divergence\%}
\title{Numerical and Symbolic Divergence}
\usage{
divergence(
  f,
  var,
  coordinates = "cartesian",
  accuracy = 2,
  stepsize = NULL,
  drop = TRUE,
  ...
)

f \%divergence\% var
}
\arguments{
\item{f}{function, expression or character array.}

\item{var}{character vector, giving the variable names with respect to which derivatives will be computed. If a named vector is provided, derivatives will be computed at that point.}

\item{coordinates}{coordinate system to use. One of: \code{cartesian}, \code{polar}, \code{spherical}, \code{cylindrical}, \code{parabolic}, \code{parabolic-cylindrical} or a character vector of scale factors for each varibale.}

\item{accuracy}{accuracy degree for numerical derivatives.}

\item{stepsize}{finite differences stepsize for numerical derivatives. Auto-optimized by default.}

\item{drop}{drop dimensions when...}

\item{...}{additinal arguments passed to \code{f}, when \code{f} is a \code{function}.}
}
\value{
divergence array.
}
\description{
Computes the divergence of functions, expressions and characters.
}
\section{Functions}{
\itemize{
\item \code{divergence}: arbitrary coordinate system

\item \code{\%divergence\%}: cartesian coordinates
}}

\examples{
# divergence of a vector field
f <- c('x^2','y^3','z^4')
divergence(f, var = c('x','y','z'))
f \%divergence\% c('x','y','z')

# numerical divergence of a vector field
f <- function(x,y,z) c(x^2, y^3, z^4)
divergence(f, var = c('x'=1,'y'=1,'z'=1))
f \%divergence\% c('x'=1,'y'=1,'z'=1)

# divergence in polar coordinates
f <- c('sqrt(r)/10','sqrt(r)')
divergence(f, var = c('r','phi'), coordinates = 'polar')

}
