% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{e2c}
\alias{e2c}
\title{Expression to Character}
\usage{
e2c(x)
}
\arguments{
\item{x}{expressions}
}
\value{
characters
}
\description{
Converts expressions to characters
}
\examples{
# convert expressions
expr <- parse(text = 'a')
e2c(expr)

# convert array of expressions
expr <- array(parse(text = 'a'), dim = c(2,2))
e2c(expr)

}
\seealso{
\code{\link{c2e}}
}
