% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate Characters and Expressions}
\usage{
evaluate(f, var)
}
\arguments{
\item{f}{array of \code{characters} or \code{expressions} to be evaluated.}

\item{var}{named vector or \code{data.frame} in which \code{f} is to be evaluated.}
}
\value{
Evaluated object. When \code{var} is a named vector, the return is an array 
with the same dimensions of \code{f}. When \code{var} is a \code{data.frame}, the
return is a \code{matrix} with columns corresponding to the entries of \code{f} and 
rows corresponding to the rows of \code{var}.
}
\description{
Evaluates an array of \code{characters} or \code{expressions}.
}
\examples{
### single evaluation
f <- array(letters[1:4], dim = c(2,2))
var <- c(a = 1, b = 2, c = 3, d = 4)
evaluate(f, var)

### vectorized evaluation
f <- array(letters[1:4], dim = c(2,2))
var <- data.frame(a = 1:3, b = 2:4, c = 3:5, d = 4:6)
evaluate(f, var)

}
\seealso{
Other utilities: 
\code{\link{c2e}()},
\code{\link{e2c}()},
\code{\link{wrap}()}
}
\concept{utilities}
