% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%div\%}
\alias{\%div\%}
\title{Numerical and Symbolic Division}
\usage{
x \%div\% y
}
\arguments{
\item{x}{character or numeric array.}

\item{y}{character or numeric array.}
}
\value{
character or numeric array.
}
\description{
Divide numeric or character arrays element by element.
}
\examples{
# div vector 
x <- c("a+1","b+2")
x \%div\% x

# div matrix 
x <- matrix(letters[1:4], ncol = 2)
x \%div\% x

# div array
x <- array(letters[1:12], dim = c(2,2,3))
y <- array(1:12, dim = c(2,2,3))
x \%div\% x
y \%div\% y
x \%div\% y 

}
