% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%dot\%}
\alias{\%dot\%}
\title{Numerical and Symbolic Dot Product}
\usage{
x \%dot\% y
}
\arguments{
\item{x}{character or numeric array.}

\item{y}{character or numeric array.}
}
\value{
character or numeric array.
}
\description{
Computes the inner product on the last dimensions of two character or numeric arrays.
}
\examples{
# inner product 
x <- array(1:12, dim = c(3,4))
x \%dot\% x

# inner product on last dimensions 
x <- array(1:24, dim = c(3,2,4))
y <- array(letters[1:8], dim = c(2,4))
x \%dot\% y
y \%dot\% x

}
