% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%prod\%}
\alias{\%prod\%}
\title{Numerical and Symbolic Product}
\usage{
x \%prod\% y
}
\arguments{
\item{x}{\code{numeric} or \code{character} array.}

\item{y}{\code{numeric} or \code{character} array.}
}
\value{
\code{array}.
}
\description{
Elementwise product of \code{numeric} or \code{character} arrays.
}
\examples{
### vector 
x <- c("a+1","b+2")
x \%prod\% x

### matrix 
x <- matrix(letters[1:4], ncol = 2)
x \%prod\% x

### array
x <- array(letters[1:12], dim = c(2,2,3))
y <- array(1:12, dim = c(2,2,3))
x \%prod\% y

}
