% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%sum\%}
\alias{\%sum\%}
\title{Numerical and Symbolic Sum}
\usage{
x \%sum\% y
}
\arguments{
\item{x}{\code{numeric} or \code{character} array.}

\item{y}{\code{numeric} or \code{character} array.}
}
\value{
\code{array}.
}
\description{
Elementwise sum of \code{numeric} or \code{character} arrays.
}
\examples{
### vector 
x <- c("a+1","b+2")
x \%sum\% x

### matrix 
x <- matrix(letters[1:4], ncol = 2)
x \%sum\% x

### array
x <- array(letters[1:12], dim = c(2,2,3))
y <- array(1:12, dim = c(2,2,3))
x \%sum\% y

}
\references{
Guidotti E (2022). "calculus: High-Dimensional Numerical and Symbolic Calculus in R." Journal of Statistical Software, 104(5), 1-37. \doi{10.18637/jss.v104.i05}
}
\seealso{
Other basic arithmetic: 
\code{\link{\%diff\%}()},
\code{\link{\%div\%}()},
\code{\link{\%dot\%}()},
\code{\link{\%inner\%}()},
\code{\link{\%kronecker\%}()},
\code{\link{\%outer\%}()},
\code{\link{\%prod\%}()}
}
\concept{basic arithmetic}
