% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.R
\name{inverse}
\alias{inverse}
\title{Numeric and Symbolic Matrix Inverse}
\usage{
inverse(x)
}
\arguments{
\item{x}{numeric or character matrix.}
}
\value{
numeric or character matrix.
}
\description{
Calculates the inverse of a matrix.
}
\examples{

# numeric matrix
x <- matrix(1:4, nrow = 2)
inverse(x)

# symbolic matrix
x <- matrix(letters[1:4], nrow = 2)
inverse(x)

}
