% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronecker.R
\name{kronecker}
\alias{kronecker}
\title{Generalized Kronecker Delta}
\usage{
kronecker(n, p = 1)
}
\arguments{
\item{n}{number of elements for each dimension.}

\item{p}{order of the generalized Kronecker delta, \code{p=1} for the standard Kronecker delta.}
}
\value{
array representing the generalized Kronecker delta tensor.
}
\description{
Computes the Generalized Kronecker Delta.
}
\examples{
# Kronecker delta 3x3
kronecker(3)

# generalized Kronecker delta 3x3 of order 2 -> 3x3 x 3x3
kronecker(3, p = 2)  

}
