% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levicivita.R
\name{levicivita}
\alias{levicivita}
\title{Levi-Civita Symbol}
\usage{
levicivita(n)
}
\arguments{
\item{n}{dimension}
}
\value{
array representing the Levi-Civita tensor.
}
\description{
Computes the Levi-Civita totally antisymmetric tensor.
}
\examples{
# Levi-Civita tensor in 2-d
levicivita(2)

# Levi-Civita tensor in 3-d
levicivita(3)

}
