% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derivative & definiteIntegral.R
\name{derivative}
\alias{derivative}
\title{Derivative function}
\usage{
derivative(expression = NULL, Point, variable = "X")
}
\arguments{
\item{expression}{A string contatining the function to differentiate.}

\item{Point}{Value at wich to differentiate.}

\item{variable}{a string containing the variable used in @param expression, Make sure the variable
string does not show anywhere else in @param expression (Ex: variable "x" in "exp(x)") To avoid this,
defaults to "X".}
}
\description{
Calculates the derivative of a function at a given point.
}
\examples{
derivative("sin(X)*exp(X)",12)
}
