% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{e2c}
\alias{e2c}
\title{Expressions to Characters}
\usage{
e2c(x)
}
\arguments{
\item{x}{\code{expression}.}
}
\value{
\code{character}.
}
\description{
Converts \code{expressions} to \code{characters}.
}
\examples{
### convert expressions
expr <- parse(text = "a")
e2c(expr)

### convert array of expressions
expr <- array(parse(text = "a"), dim = c(2,2))
e2c(expr)

}
\references{
Guidotti, E. (2020). "calculus: High dimensional numerical and symbolic calculus in R". \url{https://arxiv.org/abs/2101.00086}
}
\seealso{
Other utilities: 
\code{\link{c2e}()},
\code{\link{evaluate}()},
\code{\link{wrap}()}
}
\concept{utilities}
