% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%inner\%}
\alias{\%inner\%}
\title{Numerical and Symbolic Inner Product}
\usage{
x \%inner\% y
}
\arguments{
\item{x}{character or numeric array.}

\item{y}{character or numeric array.}
}
\value{
character or numeric.
}
\description{
Computes the inner product of two character or numeric arrays.
}
\examples{
# numeric inner product 
x <- array(1:12, dim = c(3,4))
x \%inner\% x

# symbolic inner product 
x <- array(letters[1:12], dim = c(3,4))
x \%inner\% x

}
