% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%kronecker\%}
\alias{\%kronecker\%}
\title{Numerical and Symbolic Kronecker Product}
\usage{
x \%kronecker\% y
}
\arguments{
\item{x}{character or numeric array.}

\item{y}{character or numeric array.}
}
\value{
character or numeric array.
}
\description{
Computes the Kronecker product of two character or numeric arrays.
}
\examples{
# numeric Kronecker product 
c(1,2) \%kronecker\% c(2,3)

# symbolic Kronecker product 
array(1:4, dim = c(2,2)) \%kronecker\% c('c','d')

}
