% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%outer\%}
\alias{\%outer\%}
\title{Numerical and Symbolic Outer Product}
\usage{
x \%outer\% y
}
\arguments{
\item{x}{character or numeric array.}

\item{y}{character or numeric array.}
}
\value{
character or numeric array.
}
\description{
Computes the outer product of two character or numeric arrays.
}
\examples{
# numeric outer product 
c(1,2) \%outer\% c(2,3)

# symbolic outer product 
c('a','b') \%outer\% c('c','d') \%outer\% c('e','f')

}
