% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%prod\%}
\alias{\%prod\%}
\title{Numerical and Symbolic Product}
\usage{
x \%prod\% y
}
\arguments{
\item{x}{character or numeric array.}

\item{y}{character or numeric array.}
}
\value{
character or numeric array.
}
\description{
Multiplies numeric or character arrays element by element.
}
\examples{
# prod vector 
x <- c("a+1","b+2")
x \%prod\% x

# prod matrix 
x <- matrix(letters[1:4], ncol = 2)
x \%prod\% x

# prod array
x <- array(letters[1:12], dim = c(2,2,3))
y <- array(1:12, dim = c(2,2,3))
x \%prod\% x
y \%prod\% y
x \%prod\% y

}
