% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%sum\%}
\alias{\%sum\%}
\title{Numerical and Symbolic Sum}
\usage{
x \%sum\% y
}
\arguments{
\item{x}{character or numeric array.}

\item{y}{character or numeric array.}
}
\value{
character or numeric array.
}
\description{
Adds numeric or character arrays element by element.
}
\examples{
# sum vector 
x <- c("a+1","b+2")
x \%sum\% x

# sum matrix 
x <- matrix(letters[1:4], ncol = 2)
x \%sum\% x

# sum array
x <- array(letters[1:12], dim = c(2,2,3))
y <- array(1:12, dim = c(2,2,3))
x \%sum\% x
y \%sum\% y
x \%sum\% y

}
