% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{index}
\alias{index}
\alias{index<-}
\title{Einstein Notation Indices}
\usage{
index(x)

index(x) <- value
}
\arguments{
\item{x}{array.}

\item{value}{vector of indices.}
}
\value{
array indices.
}
\description{
Get and set indices: names of the array's dimensions. See also \code{\link{einstein}}.
}
\section{Functions}{
\itemize{
\item \code{index}: get indices.

\item \code{index<-}: set indices.
}}

\examples{
# define array
a <- array(1, dim = c(1,3,2))

# get indices
index(a)

# set indices
index(a) <- c('i', 'j', 'k')

# get indices
index(a)

# dimensions
dim(a)

}
\seealso{
\code{\link{einstein}}, \code{\link[base]{dim}}
}
