% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wrap}
\alias{wrap}
\title{Wrap Character}
\usage{
wrap(x)
}
\arguments{
\item{x}{characters}
}
\value{
wrapped characters.
}
\description{
Wraps characters in round brackets.
}
\details{
Characters are automatically wrapped when performing basic symbolic operations to prevent unwanted results. E.g.: 
\deqn{a+b * c+d} 
instead of 
\deqn{(a+b) * (c+d)}
To disable this behaviour run \code{options(calculus.auto.wrap = FALSE)}.
}
\examples{
# wrap characters
wrap('a+b')

# wrap array of characters
wrap(array(letters[1:9], dim = c(3,3)))

}
